Move the PublicationPage class to the publications package
