+++ packages/framework/tests/Feature/Actions/PublicationPageCompilerTest.php
@@ -8,8 +8,8 @@ use function file_get_contents;
 use function file_put_contents;
 use Hyde\Framework\Exceptions\FileNotFoundException;
 use Hyde\Hyde;
-use Hyde\Pages\PublicationPage;
 use Hyde\Publications\Actions\PublicationPageCompiler;
+use Hyde\Publications\Models\PublicationPage;
 use Hyde\Publications\Models\PublicationType;
 use Hyde\Support\Facades\Render;
 use Hyde\Testing\TestCase;
