+++ packages/framework/tests/Feature/PublicationTypeTest.php
@@ -7,9 +7,9 @@ namespace Hyde\Framework\Testing\Feature;
 use function array_merge;
 use function array_reverse;
 use Hyde\Hyde;
-use Hyde\Pages\PublicationPage;
 use Hyde\Publications\Models\PublicationFieldDefinition;
 use Hyde\Publications\Models\PublicationListPage;
+use Hyde\Publications\Models\PublicationPage;
 use Hyde\Publications\Models\PublicationType;
 use Hyde\Publications\Paginator;
 use Hyde\Publications\PublicationService;
