+++ packages/publications/resources/views/publication_detail.blade.php
@@ -2,7 +2,7 @@
 @section('content')
     <main id="content" class="mx-auto max-w-7xl py-16 px-8">
         <article class="prose dark:prose-invert">
-            @php/** @var \Hyde\Pages\PublicationPage $publication*/@endphp
+            @php/** @var \Hyde\Publications\Models\PublicationPage $publication*/@endphp
             <h1>{{ $publication->title }}</h1>
             <p>
                 {{ $publication->markdown }}
