+++ packages/publications/src/Models/PublicationPage.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Pages;
+namespace Hyde\Publications\Models;
 
 use Hyde\Markdown\Models\FrontMatter;
 use Hyde\Markdown\Models\Markdown;
+use Hyde\Pages\Concerns;
 use Hyde\Pages\Contracts\DynamicPage;
 use Hyde\Publications\Actions\PublicationPageCompiler;
-use Hyde\Publications\Models\PublicationType;
 use function str_starts_with;
 
 /**
